Updated code to take command line arguments. Added checks to ensure input is received. Added a sample run as comments.

+++ run_ensembles.sh
@@ -1,10 +1,42 @@
-# !/bin/bash
+# Example command:
+# ./test.sh -o path/to/output/dir -n 16
 
-path=/global/scratch/zexuanxu/lbnl-amanzi-emsemble-sim/simulation_runs
-for n in {0..15}; do
-  cd ${path}/sim${n}
-  rm -f run_amanzi.sh
-  cp ../shared_files/run_amanzi.sh run_sim${n}.sh
-  sbatch run_sim${n}.sh
+# -o: output path
+# -n: number of simulations
+
+##########################
+# GET INPUTS
+##########################
+while getopts o:n:s: flag
+do
+    case "${flag}" in
+        o) output_path=$(readlink -f ${OPTARG});;
+        n) num_sims=${OPTARG};;
+    esac
+done
+
+# checks if variable is empty
+if [ -z "$output_path" ]; then
+    printf "ERROR: Argument for output path (-o) not found!\n"
+    exit 1
+fi
+
+if [ -z "$num_sims" ]; then
+    printf "ERROR: Argument for number of simulations (-n) not found!\n"
+    exit 1
+fi
+
+echo "Output path => $output_path"
+echo "Number of simulations => $num_sims"
+
+##########################
+# RUN ALL SIMULATIONS
+##########################
+for (( n=0; n<$num_sims; n++ ));
+do
+    cd ${output_path}/sim${n}
+    rm -f run_amanzi.sh
+    cp ${output_path}/shared_files/run_amanzi.sh run_sim${n}.sh
+    sbatch run_sim${n}.sh
 
 done
